+++ yii2-adapter/legacy/elements/Tag.php
@@ -9,7 +9,6 @@ namespace craft\elements;
 
 use Craft;
 use craft\elements\conditions\tags\TagCondition;
-use craft\elements\db\TagQuery;
 use craft\gql\interfaces\elements\Tag as TagInterface;
 use craft\helpers\Db;
 use craft\models\TagGroup;
@@ -21,6 +20,7 @@ use CraftCms\Cms\FieldLayout\FieldLayout;
 use CraftCms\Cms\Twig\Attributes\AllowedInSandbox;
 use CraftCms\Cms\User\Elements\User;
 use CraftCms\RulesetValidation\Attributes\Ruleset;
+use CraftCms\Yii2Adapter\Element\Queries\TagQuery;
 use CraftCms\Yii2Adapter\Validation\LegacyElementRules;
 use GraphQL\Type\Definition\Type;
 use yii\base\InvalidConfigException;
@@ -108,7 +108,7 @@ class Tag extends Element
      */
     public static function find(): TagQuery
     {
-        return new TagQuery(static::class);
+        return new TagQuery();
     }
 
     /**